# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Legal Research Platform API with AI capabilities
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: uploadedBy
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
...
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a single document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Document:
      type: object
      required: [id, filename, fileType, fileSize, uploadedAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        content:
          type: ["string", "null"]
        keywords:
          type: ["string", "null"]
        uploadedById:
          type: ["integer", "null"]
        uploadedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, supervisor, viewer]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        role:
          type: string
          enum: [owner, supervisor, viewer]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        role:
          type: string
          enum: [owner, supervisor, viewer]
...