# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Legal Research Platform API with AI capabilities
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  role?: UserUpdateRole;
}